+++ composer.json
@@ -14,6 +14,8 @@
         "php": ">=5.2.0"
     },
     "autoload": {
-        "classmap": ["S3.php"]
+        "psr-4": {
+            "Tpyo\\AmazonS3\\": "src/"
+        }
     }
 }
